requirements.txt
numpy
vamphost
librosa
youtube-dl
